+++ src/main/java/com/example/ecommerce_api/Models/DTOs/CartItemDTO.java
@@ -9,10 +9,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CartItemDTO {
-    private Long id;
+    private Integer id;
     private ProductDTO product_item;
 
-    public CartItemDTO(Long id, ProductDTO product_response){
+    public CartItemDTO(Integer id, ProductDTO product_response){
         this.id = id;
         product_item = product_response;
     }
