+++ src/main/java/com/example/ecommerce_api/Repositories/CartItemRepository.java
@@ -4,5 +4,5 @@ import com.example.ecommerce_api.Models.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 }
